+++ src/mxmake/tests/test_templates.py
@@ -672,8 +672,10 @@ class TestTemplates(testing.RenderTestCase):
                 MXENV_TARGET:=$(SENTINEL_FOLDER)/mxenv.sentinel
                 $(MXENV_TARGET): $(SENTINEL)
                 ifeq ("$(VENV_ENABLED)", "true")
+                ifeq ("$(VENV_CREATE)", "true")
                 	@echo "Setup Python Virtual Environment under '$(VENV_FOLDER)'"
                 	@$(PYTHON_BIN) -m venv $(VENV_FOLDER)
+                endif
                 endif
                 	@$(MXENV_PATH)pip install -U pip setuptools wheel
                 	@$(MXENV_PATH)pip install -U $(MXDEV)
@@ -690,7 +692,9 @@ class TestTemplates(testing.RenderTestCase):
                 .PHONY: mxenv-clean
                 mxenv-clean: mxenv-dirty
                 ifeq ("$(VENV_ENABLED)", "true")
+                ifeq ("$(VENV_CREATE)", "true")
                 	@rm -rf $(VENV_FOLDER)
+                endif
                 else
                 	@$(MXENV_PATH)pip uninstall -y $(MXDEV)
                 	@$(MXENV_PATH)pip uninstall -y $(MXMAKE)